requirements.txt
streamlit
python-docx
google-generativeai
requests